% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{add_theme}
\alias{add_theme}
\title{themes}
\usage{
add_theme(wb, coordinates, theme = list(color = "Reds"))
}
\description{
themes
}
\author{
Serkan Korkmaz \href{mailto:serkor1@duck.com}{serkor1@duck.com}
}
