% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook.R
\name{add_worksheet}
\alias{add_worksheet}
\title{.add_worksheet}
\usage{
add_worksheet(wb = NULL, list_names = NULL)
}
\description{
.add_worksheet
}
\author{
Serkan Korkmaz \href{mailto:serkor1@duck.com}{serkor1@duck.com}
}
\keyword{internal}
