% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{as_workbook_data}
\alias{as_workbook_data}
\title{as_workbook_data}
\usage{
as_workbook_data(DT, sheet, colum, row)
}
\arguments{
\item{DT}{a data.table}

\item{sheet}{character.}

\item{colum}{characte. The column split within
each sheet.}

\item{row}{character. The row split}
}
\description{
This function prepares the data
that is fed into to generate workbook funciton
}
\author{
Serkan Korkmaz \href{mailto:serkor1@duck.com}{serkor1@duck.com}
}
