% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{data_backend}
\alias{data_backend}
\title{data_coordinates}
\usage{
data_backend(list, theme = list(compact = TRUE, combine = FALSE))
}
\description{
data_coordinates
}
\author{
Serkan Korkmaz \href{mailto:serkor1@duck.com}{serkor1@duck.com}
}
