% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{data_coordinates}
\alias{data_coordinates}
\title{data_coordinates}
\usage{
data_coordinates(list, theme = list(compact = TRUE, combine = FALSE))
}
\description{
data_coordinates
}
\author{
Serkan Korkmaz \href{mailto:serkor1@duck.com}{serkor1@duck.com}
}
