% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.R
\name{.grouping_row}
\alias{.grouping_row}
\title{data_wb_backend}
\usage{
.grouping_row(wb, wb_backend)
}
\description{
data_wb_backend
}
\author{
Serkan Korkmaz \href{mailto:serkor1@duck.com}{serkor1@duck.com}
}
