% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.pkg_error}
\alias{.pkg_error}
\title{function information}
\usage{
.pkg_error(header, param, description)
}
\arguments{
\item{header}{a character string with
the warning hader}

\item{param}{a string with the parameters
causing the error.}

\item{description}{a description of the error.}
}
\description{
function information
}
