% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.pkg_warning}
\alias{.pkg_warning}
\title{function information}
\usage{
.pkg_warning(header, body)
}
\arguments{
\item{header}{a character string with the
warning header.}

\item{body}{a character string with
warning body.}
}
\description{
function information
}
