% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_workbook.R
\name{generate_workbook}
\alias{generate_workbook}
\title{generate_workbook}
\usage{
generate_workbook(
  file = "workbook.xlsx",
  list = NULL,
  creator = NULL,
  title = NULL,
  subject = NULL,
  category = NULL,
  theme = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{file}{A character string naming an xlsx file}

\item{list}{A list of data.tables, can be nested. Or created using the as_workbook_data function}

\item{creator}{Creator of the workbook (your name). Defaults to login username}

\item{title}{Workbook properties title}

\item{subject}{Workbook properties subject}

\item{category}{Workbook properties category}

\item{theme}{A named list of parameters}

\item{overwrite}{If \code{TRUE}, overwrite any existing file.}
}
\value{
A workbook
}
\description{
Generate a Workbook which is also stored in \code{file}.
}
\examples{
# script start; ####

# gc
invisible(
  {
    rm(list = ls()); gc()
  }
)

# load libraries
library(data.table)
library(workbookR)

# 1) prepare the data;
mtcars <- as.data.table(
  mtcars
)

# 1.1) convert relevant
# variables that determines
# the split of the data
mtcars[
  ,
  `:=`(
    cyl_   = paste(cyl, 'cylinder(s)'),
    gear_  = paste(gear, 'Gears'),
    am_    = fcase(
      default = 'Automatic',
      am == 1, 'Manual'
    )
  )
  ,
]


# 2) generate workbook_data
# using as_workbook_data
workbook_data <- as_workbook_data(
  DT    = mtcars,
  # splitting by sheets
  sheet = 'am_',
  # splitting each data within 
  # the sheets by number of
  # gears
  colum = 'gear_',
  # Splitting all data
  # between number of cylinders
  row   = 'cyl_'
)

# 3) generate the workbook
# and store in file
wb <- generate_workbook(
  file = 'workbook.xlsx',
  list = workbook_data
)


# end of script; ####
}
