% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{get_caption_coords}
\alias{get_caption_coords}
\title{data_coordinates}
\usage{
get_caption_coords(DT, caption, subcaption, adjust)
}
\description{
data_coordinates
}
\author{
Serkan Korkmaz \href{mailto:serkor1@duck.com}{serkor1@duck.com}
}
