% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{get_table_coords}
\alias{get_table_coords}
\title{function information}
\usage{
get_table_coords(x_ = 3, y_ = 10, DT, grouping_name = "Placeholder")
}
\arguments{
\item{x_}{the starting x coordinate
for the data. Has to be an integer}

\item{y_}{the starting y coordinate
for the data. has to be an integer}

\item{DT}{the data.table that needs
to be written to the workbook}
}
\description{
function information
}
