% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbookR-package.R
\docType{package}
\name{workbookR-package}
\alias{workbookR}
\alias{workbookR-package}
\title{workbookR: Create Woorkbooks with a high level Openxlsx API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A high level openxlsx wrapper for tables
}
\author{
\strong{Maintainer}: Serkan Korkmaz \email{info@1903analytics.ai} (\href{https://orcid.org/0000-0002-5052-0982}{ORCID})

}
\keyword{internal}
