% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_workbook.R
\name{generate_workbook}
\alias{generate_workbook}
\title{generate_workbook}
\usage{
generate_workbook(
  file = "workbook.xlsx",
  list = NULL,
  creator = NULL,
  title = NULL,
  subject = NULL,
  category = NULL,
  overwrite = TRUE
)
}
\value{
stores and returns a workbook in the specified
file path
}
\description{
This function generates and collects
the workvook
}
