% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_workbook.R
\name{generate_workbook}
\alias{generate_workbook}
\title{generate_workbook}
\usage{
generate_workbook(
  file = "workbook.xlsx",
  list = NULL,
  creator = NULL,
  title = NULL,
  subject = NULL,
  category = NULL,
  theme = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{file}{A character string naming an xlsx file}

\item{creator}{Creator of the workbook (your name). Defaults to login username}

\item{title}{Workbook properties title}

\item{subject}{Workbook properties subject}

\item{category}{Workbook properties category}

\item{theme}{A named list of parameters}

\item{overwrite}{If \code{TRUE}, overwrite any existing file.}
}
\value{
A workbook
}
\description{
Generate a Workbook which is also stored in \code{file}.
}
