% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.R
\name{table_headers}
\alias{table_headers}
\title{data_coordinates}
\usage{
table_headers(wb, coordinates)
}
\description{
data_coordinates
}
\author{
Serkan Korkmaz \href{mailto:serkor1@duck.com}{serkor1@duck.com}
}
